requirements.txt
requests
pytest
kafka-python
psycopg2-binary
